Starting off with math for exponents. First, I need to be able to add constants together (for adding exponents).
Added an "AddConstant" and "SubtractConstant" function

Also added an "IsANumber" function to determine if a term is numeric. This function assumes the term is not complex.

Finally, modifying the "CancelSubtraction" function to employ the subtraction of constants.

Unresolved: sign change on the subtraction operation -- what do I do if the sign changes?